Encapsulate canonical id map
