+++ crates/polars-plan/src/plans/optimizer/cse/mod.rs
@@ -1,7 +1,7 @@
 mod cache_states;
+mod canonical_ir;
 mod csee;
 pub mod cspe;
-mod interner;
 
 pub(crate) use cache_states::set_cache_states;
 pub(super) use csee::CommonSubExprOptimizer;
